+++ app/controllers/backoffice/services/bundles/publishes_controller.rb
@@ -5,10 +5,11 @@ class Backoffice::Services::Bundles::PublishesController < Backoffice::Applicati
 
   def create
     if Bundle::Publish.call(@bundle)
+      flash[:notice] = "Bundle published successfully"
       redirect_to backoffice_service_path(@service)
     else
       flash[:alert] =
-        "Bundle not published, errors: " +
+        "Bundle cannot be published. Please ensure your form is properly completed. " +
           "#{@bundle.errors.messages.each.map { |k, v| "The field #{k} #{v.join(", ")}" }.join(", ")}"
       redirect_to edit_backoffice_service_bundle_path(@service, @bundle)
     end