+++ app/controllers/services/logos_controller.rb
@@ -3,7 +3,7 @@
 class Services::LogosController < ApplicationController
   def show
     @service = Service.friendly.find(params[:service_id])
-    authorize(ServiceContext.new(@service, false))
+    authorize(ServiceContext.new(@service, params.key?(:from) && params[:from] == "backoffice_service"))
 
     if @service.logo.attached? && @service.logo.variable?
       redirect_to @service.logo.variant(resize: "84x84"), allow_other_host: false