+++ spec/services/service/unpublish_spec.rb
@@ -2,15 +2,15 @@
 
 require "rails_helper"
 
-RSpec.describe Service::Draft, backend: true do
+RSpec.describe Service::Unpublish, backend: true do
   it "draft service" do
     service = create(:service)
     offer = create(:offer, service: service)
     service.reload
     described_class.call(service)
 
-    expect(service.reload).to be_draft
-    expect(offer.reload).to_not be_draft
+    expect(service.reload).to be_unpublished
+    expect(offer.reload).to_not be_unpublished
   end
 
   context "#bundled_offers" do